Revert "Fix compiler warning."
This reverts commit d779d03d91fde17f4b00f508ceaed4583e328b41.

+++ trunk/src/jsformatter.cpp
@@ -80,7 +80,7 @@ DLLAPI const char *JSFGetDebugOutput(JSFormatter *jsf)
 	return ((RealJSFormatter *)jsf)->GetDebugOutput();
 }
 
-DLLAPI long JSFGetFormattedLine(JSFormatter *jsf, long originalLine)
+DLLAPI int JSFGetFormattedLine(JSFormatter *jsf, int originalLine)
 {
 	return ((RealJSFormatter *)jsf)->GetFormattedLine(originalLine);
 }
